requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
motor==3.6.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.9.2
email-validator==2.2.0

